requirements.txt
transformers
datasets
pytorch-lightning
scikit-learn
pandas
wandb
rich
typer
loguru
black
isort
flake8
flake8-bugbear
mypy
ipywidgets
pre-commit
bitsandbytes
pytorch_optimizer
